requirements.txt
Flask>=3.0.0,<4.0.0
PyYAML>=6.0
gpiozero>=2.0
adafruit-circuitpython-pca9685>=3.4.15
adafruit-circuitpython-motor>=3.4.8
smbus2>=0.4.3
adafruit-blinka>=8.0.0

evdev>=1.7.0

pexpect>=4.8.0
opencv-python-headless
numpy

psutil>=5.9.0

ultralytics>=8.0.0
